+++ src/qnethernet/QNMDNS.h
@@ -14,9 +14,8 @@
 #include <cstddef>
 #include <cstdint>
 #include <limits>
+#include <string>
 #include <vector>
-
-#include <WString.h>
 
 #include "lwip/apps/mdns.h"
 #include "lwip/netif.h"
@@ -95,7 +94,7 @@ class MDNSClass final {
   //
   // If this returns false and there was an error then errno will be set.
   bool addService(const char* type, const char* protocol, uint16_t port,
-                  std::vector<String> (*getTXTFunc)());
+                  std::vector<std::string> (*getTXTFunc)());
 
   // Adds a service. The protocol will be set to "_udp" for anything other than
   // "_tcp". The strings should have a "_" prefix.
@@ -109,7 +108,7 @@ class MDNSClass final {
   // If this returns false and there was an error then errno will be set.
   bool addService(const char* name, const char* type,
                   const char* protocol, uint16_t port,
-                  std::vector<String> (*getTXTFunc)());
+                  std::vector<std::string> (*getTXTFunc)());
 
   // Removes a service. The host name is used as the service name. This will
   // return whether the service was removed.
@@ -142,7 +141,7 @@ class MDNSClass final {
 
     void set(bool valid, const char* name, const char* type,
              enum mdns_sd_proto proto, uint16_t port,
-             std::vector<String> (*getTXTFunc)());
+             std::vector<std::string> (*getTXTFunc)());
     ATTRIBUTE_NODISCARD
     bool equals(bool valid, const char* name, const char* type,
                 enum mdns_sd_proto proto, uint16_t port) const;
@@ -156,7 +155,7 @@ class MDNSClass final {
     char type_[MDNS_LABEL_MAXLEN + 1];
     enum mdns_sd_proto proto_ = mdns_sd_proto::DNSSD_PROTO_UDP;
     uint16_t port_ = 0;
-    std::vector<String> (*getTXTFunc_)() = nullptr;
+    std::vector<std::string> (*getTXTFunc_)() = nullptr;
   };
 
   MDNSClass() = default;